backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlmodel==0.0.31
asyncpg==0.31.0
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4
httpx==0.28.1
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
slowapi==0.1.9
websockets==16.0
Pillow==12.1.0
bleach==6.3.0
tinycss2==1.5.1
aiofiles==25.1.0
alembic==1.18.1

pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
redis>=5.2.0
